Change Text.x and Text.y to floats

+++ engine/core/text.py
@@ -20,8 +20,8 @@ class Text(Entity):
 
     def __init__(
         self,
-        x: int = 0,
-        y: int = 0,
+        x: float = 0,
+        y: float = 0,
         text: str = "",
         size: int = 50,
         font: Optional[str] = None,
@@ -36,8 +36,8 @@ class Text(Entity):
         Text entities behave like any other entity, but they render a string of text instead of an image or colored box.
 
         Args:
-            x (int): X position. Defaults to 0.
-            y (int): Y position. Defaults to 0.
+            x (float): X position. Defaults to 0.
+            y (float): Y position. Defaults to 0.
             text (str): The string to render. Defaults to "".
             size (int): Font size in points. Defaults to 50.
             font (Optional[str]): Path to a font file. Defaults to None (pygame's
@@ -51,8 +51,8 @@ class Text(Entity):
                 to avoid the per-frame re-render cost. Defaults to True.
         """
 
-        self.x: int = x
-        self.y: int = y
+        self.x: float = x
+        self.y: float = y
         self.text: str = text
         self.color: RGBType = color
         self.bgcolor: RGBType = bgcolor
